finance_monitor/requirements.txt

pandas==2.2.2
requests==2.31.0
yfinance==0.2.36
flask==2.3.3
gunicorn==21.2.0
matplotlib==3.8.4
python-dotenv==1.0.1
slack_sdk==3.21.3
pytrends==4.9.0
supabase-py==1.0.3
b2sdk==1.22.0
pytest==8.2.0
apscheduler==3.10.1
